Suppress MSVC 8+ deprecation warnings about standard C/C++ library functions.
Using many standard C/C++ library functions and also all POSIX functions
provided by MSVC CRT results in a deprecation warning with MSVC 8 or later.

As we do want to be able to use these functions, suppress the warnings
globally by defining _CRT_SECURE_NO_WARNINGS and _CRT_NONSTDC_NO_WARNINGS on
compiler command line: this is much simpler and less intrusive than doing it
in the headers as it can't be done in config_msvc.hpp (it is included too late
as some standard headers are already included from platform_dependent.hpp
before it) and so would require some changes to common code which we can avoid
by doing it at this level.