Upgrade Windows build system to Boost 1.38.0.
This patch upgrades Boost used in the MSW build system from 1.33.1 to
1.38.0. It is a prerequisite for the upcoming patch to use
boost::math::expm1() and log1p().

It fixes the problem with filesystem compilation (discussed here:
http://lists.nongnu.org/archive/html/lmi/2008-06/msg00054.html) by
tweaking compiler flags a bit. In other words, it doesn't yet change the
way Boost libraries are built in none of the ways also mentioned in the
above thread.