Honour XMLWRAPP_{CFLAGS,LIB} variables when not using pkg-config too.
pkg-config allows the user to predefine these variables to avoid running
checks for the library. Allow using them when not using pkg-config too for
consistency and because these variables are documented in any case in
"configure --help" output so they should work. Besides, it's more convenient
to use them instead of xmlwrapp-config script when cross-compiling.