Honour XMLWRAPP_{CFLAGS,LIB} variables when not using pkg-config too.
pkg-config allows the user to predefine these variables to avoid running
checks for the library. Allow using them when not using pkg-config too for
consistency and because these variables are documented in any case in
"configure --help" output so they should work. Besides, it's more convenient
to use them instead of xmlwrapp-config script when cross-compiling.

git-svn-id: svn://svn.sv.gnu.org/lmi/lmi/trunk@5218 8995fb05-a277-41b4-a4d2-e84fd280a9a6