Filter out disallowed mc_enum values in CensusView.
Implement the equivalent of MvcController::ConditionallyEnableItems().
Because wxDataViewChoiceRenderer only works with static choices list,
implement a custom renderer that is very similar to the generic
wxDataViewChoiceRenderer implementation, except that it determines the
list of allowed choices right before showing the editor control.

This requires passing information about the Cell (row) to the renderer
through wxVariant, so a custom rendeder is needed anyway and stock
wxDataViewChoiceRenderer couldn't be used anyway, even if it were
modified to be more flexible. Overall, it's probably better to do it
like this.