Add support for multi-dimensional tables.
Load multiple basic tables from XML file. The new queries API takes a
database_index key that is used to choose an appropriate basic table.

Old API is still provided for compatibility. It only works reasonable
with single-table files (which is the only kind we care about for
compatibility), otherwise it just picks the first table in the file.