Refactor: extract test illustration/census creation code.
Add helper classes for creating illustrations and censuses. While doing it
manually only takes a few lines every time, it's still better to avoid
repeating these few lines in so many places, so extract the code simulating
creating a new document and closing it in a reusable helper class and just use
it from elsewhere.

This makes the code shorter and also safer as there are now additional checks
ensuring that the test document is closed at the end of its test.