Minimal changes to use Boost.Filesystem v3 API.
The following changes had to be done to allow the code compilation with v3
version of the library as included in Boost 1.56:

- Replace fs::path::native_file_string() with just string().

- Take into account the change of return type of fs::path::leaf() from
  std::string to fs::path.

- Remove default_name_check() call and use of fs::native in fs::path ctor
  as neither is supported any longer. The checks would need to be readded
  if they are really needed.

- Use absolute() instead of complete() as the latter is not available at all
  in the v3 API, apparently by mistake, i.e. it was supposed to be still
  provided in it and only deprecated, but actually it is not part of boost::
  filesystem namespace at all and is only defined in filesystem3 one, at least
  with 1.49.

- Avoid using change_extension() which is, again, buggy in 1.49.

Deprecated methods are still being used, see

http://www.boost.org/doc/libs/1_56_0/libs/filesystem/doc/deprecated.html

for their list and suggested replacements. They would need to be expunged to
allow eventually compiling LMI with C++14 std::filesystem.

There are also still failures in path_utility_test, some of them due to
MSW-specific tests (e.g. relying on not being able to delete a file being
currently open), but also apparently due to the changes in behaviour of the
library between v2 and v3.