Add a local copy of wxwin.m4 file from wxWidgets.
Don't rely on "wx" environment variable being set when autogen.sh is ran, this
is not reliable and doesn't work at all if automake decides to regenerate the
files for any reason.

Instead, just make wxwin.m4 containing wxWidgets-related macros used by our
configure always available as part of our sources and set ACLOCAL_AMFLAGS to
point to it, so that it's found even when regenerating the files automatically.