Don't compile wx_checks.cpp as part of configure-based build.
The checks performed in this file are already done in configure with the only
difference that they -- intentionally -- result in a warning, not a fatal
error there. This allows to build using the system version of the library
without having to use a custom wxWidgets build just for LMI, which is quite
convenient under Unix.