Don't compile wx_checks.cpp as part of configure-based build.
The checks performed in this file are already done in configure with the only
difference that they -- intentionally -- result in a warning, not a fatal
error there. This allows to build using the system version of the library
without having to use a custom wxWidgets build just for LMI, which is quite
convenient under Unix.

git-svn-id: svn://svn.sv.gnu.org/lmi/lmi/trunk@5970 8995fb05-a277-41b4-a4d2-e84fd280a9a6