Remove the possibility to build Boost.Filesystem from configure.
This wasn't really useful and didn't work correctly in almost any environment
as the list of files in configure is specific to the Boost version 1.33 which
can't be built using g++ 4 and so we would need to make it conditional to
support both 1.33 and some later version, which just doesn't seem to be worth
it.

Simply use the system version of the library instead, just as we do for
Boost.Regex already.

git-svn-id: svn://svn.sv.gnu.org/lmi/lmi/trunk@5971 8995fb05-a277-41b4-a4d2-e84fd280a9a6