Handle only the menu open events for the main frame menus.
Connect Skeleton::UponMenuOpen() only to the menu open events for the main
frame menus instead of indiscriminately handling all the events ending in the
application object.

This ensures that we never use the possibly already destroyed main frame
pointer if another frame (with a menu) comes up during the application
shutdown, as it could happen in the automated GUI tests.