Only recreate census view columns if really necessary
Avoid always recreating all the columns, this is relatively slow and,
even worse, prevents the cache of best column widths from being ever
used. Now that the cache can be used, switching to the "Varying width"
mode is 10 times faster after the first time (and still almost twice
faster when doing it the first time).