Define global swap() for any_member<>.
The default std::swap() implementation doesn't work for swapping arbitrary
any_member<> objects as it relies on being able to assign them, but this is
impossible with any_member<> which has a non-standard assignment semantics and
only supports assigning between objects of the same type. However swapping two
any_member<> objects, of whatever types, is perfectly possible, so provide
swap() overload allowing to do this.

Notice that std::swap() can't be overloaded, hence we this swap() function
must be defined in the same namespace as any_member<> itself, i.e. the global
one.