Allow computing MemberSymbolTable member access index and reusing it.
Looking up the member by name once and then reusing the index to access it
results in a big performance gain when doing it many times in a row, e.g.
CensusView::column_value_varies_across_cells() now takes less than 30% of the
time it took before for a census with ~4000 rows.