Stop using WX_STANDARD_OPTIONS() in configure
There is really no reason to add specific --with-toolkit and --with-wxshared
options to configure when it's already possible to just point it directly to
the wx-config to use.

As for --enable-debug, it clashed with the already-defined option with the
same name while --enable-shared doesn't make any sense for lmi at all.

So WX_STANDARD_OPTIONS() was completely useless and could be removed, as well
as WX_CONVERT_STANDARD_OPTIONS_TO_WXCONFIG_FLAGS which had to be used to make
the values of these options affect the choice of wx-config.

git-svn-id: svn://svn.sv.gnu.org/lmi/lmi/trunk@6288 8995fb05-a277-41b4-a4d2-e84fd280a9a6