Make columns in PDF tables wide enough for their widest text
Fix the columns width in case the widest text of the column is wider than the
column header.

Previously, the "em" margins were only added if the column header was wider,
but they need to be added in any case because the column would look too
cramped otherwise and also because the margin is used when rendering it, which
resulted in wrong display if it wasn't taken into account when determining its
width.