Add a margin on the right hand side of the summary box
This is not the best but definitely the simplest way to improve appearance of
the "Plan Details Summary" box in the group report when using the default
headers, as without this extra margin the rightmost column is much more narrow
than the leftmost one, resulting in visibly bigger margins for the latter.

Avoid this asymmetry by adding an extra margin to the right most column.