Check that the run order is the same for all cells and the case
Generalize assert_consistency() function to check that the run order for all
cells is the same as for the case instead of checking just the first one.

Benchmarking shows that the timings of illustration generation are not
significantly affected by this change, but it catches inconsistencies which
could previously pass unnoticed.