Verify census consistency before creating a group quote report
Replace assert_consistent_run_order() with a more generic function
assert_consistency_in_context() which performs the run order check and also
the check for consistency of the fields used in the group quote report if one
is to be generated.

This ensures that the data in the group quotes header and footer is correct
for all cells.