Use a single file menu adjusted in the code
Instead of defining 3 different file menus in the XRC, and duplicating big
parts of the menu thrice, define only one menu with all the items and delete
those that are inappropriate for the current view when setting up the menu bar
in the code.

Reuse the existing Skeleton::AdjustMenus() to do it and pass an additional
pointer to the view to it to allow to decide which items to delete. The view
classes hierarchy has been modified to provide a CanBePrinted() method used
for this.