Handle focus loss in InputSequenceEntry itself
Replace wxEVT_KILL_FOCUS handlers in InputSequenceTextCtrl and
InputSequenceButton classes with a single handler for both of them in
InputSequenceEntry.

This slightly simplifies the code and also ensures that the more
correct logic in the button event handler is used for the text control
as well, so that things will work correctly if the editor can ever be
shown while the text control has focus (which will eventually happen).