Allow opening editor from keyboard in InputSequenceEntry
Handle Alt-Enter in InputSequenceTextCtrl as a shortcut for opening the editor
dialog to allow doing it from keyboard even when InputSequenceTextCtrl is
embedded in wxDataViewCtrl (as it happens when it's used in CensusView) and
TAB-bing to the button is impossible as TAB presses are intercepted by
wxDataViewCtrl to pass to the next cell.

Additionally, using Alt-Enter is simpler than using Tab,Enter even when the
control is used standalone.