Fix clang build by not defining const object without default ctor
The C++ standard requires all const objects, even empty ones, such as renderer
converter objects here, to have a user-defined default ctor and clang, unlike
gcc and MSVC, does insist on this.

On one hand, clang is correct and these classes must define (trivial) default
ctors for the code to be standard-compliant. OTOH, this is an open issue in
the C++ standard with a scheduled resolution which will allow to avoid having
a default ctor in this case, see the issue 253 at

  http://www.open-std.org/jtc1/sc22/wg21/docs/cwg_active.html#253

So, instead of littering the code with the otherwise useless trivial default
ctors, just remove "const" from the singleton object definition as it doesn't
really change anything but allows the code to build with clang too.