Pass the field constant, not string name, to field reading functions
While we only need the field name for error reporting, it's more clear and
less error-prone (because there is no danger of making a typo) to pass the
field constants instead of the name string to the functions reading the field
values.

The only remaining exception is do_read_number() which is also used to read
numbers that don't correspond to any fields (record type and length).