Make validation of max select age stricter and document it
Check that max select age is always specified for the select tables instead of
assuming that unspecified or 0 max select age is equivalent to max age as this
just needlessly complicates things and doesn't seem to be useful in practice
(unlike the case of having max select age set to max age in input for the
non-select tables, which does happen).

Also document explicitly the invariants checked by validate() and which can be
relied upon by the code called after the validation.