Implement safe and correct saving of SOA databases
Ensure we don't overwrite any existing files unless we do write the new ones
successfully, even if an error occurs while writing them. Do it using the
usual trick with writing to a temporary file and then atomically renaming it
except that it needs to be done in an even more complicated way in order to
deal with two files -- but even if this solution is rather complex and still
not perfect, there doesn't seem to be any way to simplify it without
sacrificing correctness.

This also corrects the bug with saving the data to the same file that the
database had been loaded from, update the tests to check for this.