Add helper location_info to SOA tables code
An object of this type more clearly indicates its meaning than a simple int
and it also allows to avoid repeating "at line" (and, to a lesser extent, "at
position") in so many error messages.

All occurrences of "int line_num" should be replaced with location_info later,
but for now just using it even if it means constructing it from line_num is
better than nothing.