Allow matching exception message substrings in BOOST_TEST_THROW
Add matcher class implementing the default (and previously the only possible)
behaviour and a match_substr subclass allowing to match the exception messages
only partially which is very convenient for the exceptions containing variable
parts (e.g. line and column numbers of the location of the error in the input).