Don't crash if the GUI test was aborted for any reason
The GUI test code could crash trying to close the already destroyed main
window on exit. In particular, this happened if the test was aborted for any
reason, including interfering with its execution by closing it interactively,
which could easily reproduce the crash, but it could also happen due to an
error during testing.