Ensure that the input sequence dialog is laid out only once
Making changes in this dialog and, especially noticeably, creating it
initially could result in too many calls to redo_layout() which was visibly
slow.

Fix this by adding "layout freeze count" allowing to temporarily ignore layout
calls and only perform the layout once at the end.

This brings down the time needed to open the dialog with a 35 element sequence
(atypical, but actually happening in practice) in ~800ms instead of ~4s
previously.