Allow switching skin while lmi is running [450]
This allows to dynamically change the skin used for the parameters
dialog while running the program and is especially useful when the same
test needs to be performed with several different skins as it allows to
do it without restarting.

To implement this, do the following:

Add ce_skin_name class representing all the available skin files in the
same way ce_product_name represents all the available product names.

Expose a field of this type corresponding to the default skin file name
stored in configurable_settings via PreferencesModel, so that it can be
edited in the preferences dialog.

Add the necessary code to reload the skin file to the code showing the
dialog (notice that this cannot be done in the PreferencesModel itself
as it is completely decoupled from the code using wxXmlResource).