Add averages sub-header to the group quotes report
Show the averages per $1000 for the premium columns.

The averages are currently computed on the fly instead of being taken from the
composite ledger as the totals are, as they're not already available elsewhere
and it's simple enough to compute them here.

Also rename output_table_totals() to output_aggregate_values() as this
function now outputs more than just the totals.