Put all group quote summary fields in the same vector
No real changes, just get rid of the code handling the first fixed fields
specially and simply add them to the vector used for the last fixed field and
all the custom ones already, the code dealing with the custom fields works
with the fixed ones too and there is no need to handle them specially.

Also replace an inconsistent use of emplace_back(extra_summary_field{...})
with just push_back({...}) which is exactly as efficient (even if it doesn't
really matter here), but is shorter and more readable.