Do not use input-sequence editor for 7702A amounts-paid history
The input-sequence editor is designed on the assumption that input
vectors begin on the issue date and change on policy anniversary. That
assumption holds for every field but 'Inforce7702AAmountsPaidHistory',
which often begins in a different year and changes on a different date
within each year. It is confusing to edit amounts-paid history in a GUI
that says it begins on the issue date when it does not.

Therefore, 'Inforce7702AAmountsPaidHistory' is now presented as a plain
wxTextCtrl rather than as an InputSequenceEntry control. It is still an
input sequence and can still be edited as a semicolon-delimited string.
In production, it is always imported from an admin extract file anyway,
and it should be edited only by experts and only on rare occasions, so
this change does not materially impair usability.

It would of course be possible to change the design of the editor to
embrace this unusual circumstance, but that would introduce needless
complexity, and 'Inforce7702AAmountsPaidHistory' is unique in deviating
from the general paradigm.