Remove spreadsheet file extension from Preferences dialog
It seemed like a good idea to add this to the dialog when other fields
were being added, even though it's hard to imagine any end user wanting
to change it. However, two problems were observed in testing. The minor
problem was that the leading dot might be removed, resulting in
  my_reporttsv
           ^ missing dot
and a special workaround seemed unjustified in light of the feature's
lack of usefulness. The major problem was that most end users have a
proprietary spreadsheet program that opens the TSV files lmi creates
seemlessly if its former default extension, '.xls', is used, but fails
if its new default, '.xlsx', is used instead. Thus, the most obvious
extension that would occur to end users would break something that has
worked for decades, so no choice should be offered at all in the GUI.
The cognoscenti prefer '.gnumeric', but know how to edit xml by hand
and need no GUI.