Use native paths with file and directory pickers
Convert paths in 'configurable_settings.xml' to native when loading,
and to generic when saving, so that the GUI always shows backslashes
under msw but they don't leak into the xml.

A normal standalone unit test is insufficient because the actual
behavior of boost::filesystem may differ when a DLL boundary is
crossed:
  http://lists.nongnu.org/archive/html/lmi/2016-06/msg00058.html
so, temporarily, a tiny test is included here. Given the apparent
fragility of this library, the msw-specific workaround for file and
directory pickers is platform conditional.