Improve error messages for invalid default-input file
Customization of input parameters is optional: it is all right to
specify the default-input file as an empty string or an invalid path,
causing lmi to fall back silently on built-in defaults. However,
attempting to edit a nonexistent file with "File | Default..."
elicited wx error messages such as
  (using empty path with wxDOC_SILENT doesn't make sense)
which might be unclear to end users, so now lmi issues its own
diagnostic instead.