Save configurable settings if print directory forcibly modified
If the print directory is invalid, then it is forcibly changed to the
current directory--at startup, as always. That change occurs before the
"Preferences" dialog is shown, so that dialog showed a valid directory,
which vanished when lmi was closed because the forcible change was
evanescent. Therefore, it is made permanent: any such forcible change
is forcibly saved.