Don't make TransferString() helper function template
This function is supposed to work with just controls containing multiple items
and not any arbitrary type T, so make this more manifest by requiring it to be
passed a wxItemContainerImmutable.

Unfortunately, this also requires passing the function a wxWindow, to get the
window name from it for error reporting, separately.