Improve layout of the "Names" page of group carveout skin
Align the labels vertically with their corresponding text controls.

Also put the street address on its own line and combine city, state and
ZIP controls on another one as they typically need less space
(especially the latter).

The text controls for the various entities are still not aligned
vertically because of an over long "Employee Class" string in the top
one, this could be fixed in the future by either shortening this label,
if possible, making another one artificially longer or, in the worst
case, by creating an invisible (but with space still reserved for it)
control with the same label in the two other zones too.