Switch to using pcre::regex instead of boost::regex
Drop dependency on Boost.Regex library and use PCRE-based regex class
instead.

Using PCRE-based implementation is advantageous compared to using
std::regex because it has a much better performance and also supports
more regex features, notably the "?<=" look-behind assertions not
supported by the standard class, that are used in the existing code.

With PCRE, the only regex change needed was the one replacing "\l",
which is not recognized by PCRE (nor std::regex) as a character class
shorthand, with the explicit use of "[:lower:]".

Currently PCRE is used only under POSIX systems and so the fully
functional version of test_coding_rules tool can only be built there
and a stub version, which simply outputs an error, is built by default
under MSW. This shouldn't be a problem in practice, because this tool is
only used during lmi development, which is done only on POSIX platforms.

The regex unit test has been minimally modified to build, and pass,
without Boost.Regex.

Finally, remove the last vestiges of Boost libraries from configure and
automake makefiles.