Forbid periods in product names
The modified unit test shows an example of the confusion that might
ensue from the laxness of the original code. This line:
  product_database::initialize("sample.database")
happened to read the file that it appears to read, by accident only.
In fact, the extension was replaced, resulting in "sample.product",
and the database indicated in that product file was opened. A line
like this:
  product_database::initialize("sample2.database")
would have read the same database file, although the file that its
argument seems to name does not actually exist: it would have read
"sample2.product", which uses "sample.database" without a "2".