Add unit test for class-instance caching
The speed test shows little benefit from caching. That is not
astonishing because the class it tests merely reads a file, without
the costly processing that motivated caching in the first place.
However, if this line:
        std::time_t const write_time = fs::last_write_time(filename);
in the caching implementation is experimentally replaced with
        std::time_t const write_time = 0;
then caching improves the speed by several orders of magnitude.