Fix harmless "unused parameter" warnings in Boost with clang
Disable -Wunused-parameter warning (implicitly enabled by -Wextra)
in Boost.Optional code when building with clang 3.8. This warning is
absolutely harmless but still results in a build error because of
-Werror.

Notice that for some unfathomable reason g++ does not give this
warning, although it looks like it should, too. But as it doesn't,
there is no need to disable it for it and we do it only for clang.