Improve typedef-names for type retrieved from cache
'value_ptr_type' and 'ptr_type' named the same type, which is the type
retrieved from the cache, so 'retrieved_type' is more descriptive.

Introduce the replacement typedef-name by alias-declarations because
that is more modern and readable. In class 'cache_file_reads', make
typedefs private because there is no reason for them to be public.