Adapt 'product_file_test' to recent database caching changes
Class DBDictionary used to be a singleton, but now it's not--it has a
normal ctor instead, as a consequence of recent changes that moved its
caching to an auxiliary class. Neither the old nor the new caching has
any effect on the speed measured by this unit test, which calls a
lower-level function to load a database from file imperatively.