Make WriteSampleDBFile() and WriteProprietaryDBFiles() static
These member functions should be static, as their analogues for other
product files are. Those analogues generally call non-static member
functions on a dummy object, as WriteProprietaryDBFiles()'s default
implementation now does. Five thousand lines of proprietary legacy code
for this class happen not to do that, so, as a temporary expedient,
in that code:
 - rename WriteProprietaryDBFiles() to Nyarlathotep()
 - copy and paste these two lines:
     void DBDictionary::WriteProprietaryDBFiles()
     {DBDictionary z; z.Nyarlathotep();}