Serve cached file reads as non-const, rather than const, shared_ptr
Remove const_cast from the unit tests. Although it was legal, it could
too easily have become illegal:
  http://lists.nongnu.org/archive/html/lmi/2016-08/msg00005.html

Add a new accessor
  DBDictionary const& product_database::db() const
and use it everywhere the shared_ptr member had been used (except in
unit tests, which aren't production code)--so that the cached database
is normally accessed through a const API, as is generally desirable:
  http://lists.nongnu.org/archive/html/lmi/2016-08/msg00006.html
Conveniently, that causes this function:
  product_database::entity_from_key(e_database_key) const
to choose the const overload of DBDictionary::datum(), when otherwise
overload resolution would prefer the non-const datum(), which is
inaccessible in this const member function. Curiously, this accessor
is private, only because there is no present need for any other class
to use it.