Overwrite 'msw_generic.make' with contents of 'msw_cygwin.make'
The original 'msw_generic.make' is obsolete, and any attempt to use it
produces an error message. It will be repurposed for cross building.
This initial step sets its contents to match 'msw_cygwin.make', which
is close to what will be needed.

A new name like 'msw_linux.make' might have been used, but the makefile
may also be used with bsd.