Make unit test more robust
When a file does not exist, boost::filesystem::last_write_time throws an
exception whose what() string includes "File not found" on GNU/Linux,
but "cannot find the file specified" on msw. It seems silly to require
either of those when another OS might say something different. In this
unit test, it is sufficient that an exception is thrown and its what()
string contains the name of the nonexistent file.

This change makes this unit test work at least as well on GNU/Linux as
it already did elsewhere, as nongnu.org requests.