Abandon xml rate table experiment
Although this was an interesting idea, we never used it in practice.

SOA has provided tables in several formats, including
 - flat text, which lmi uses and its 'rate_table_tool' supports;
 - binary, which 'rate_table_tool' supports;
 - xml, in a non-free format incompatible with the one removed here;
 - several spreadsheet formats, all of which (astonishingly, even
   their so-called CSV) are incompatible with gnumeric and libreoffice.
Compared to this or any xml format, the binary format that we use for
production is more compact and has better runtime performance, while
the text format that we use for maintenance is much easier for humans
to read and edit.