Simplify conditionals using certain object-like macros
LMI_RINT_AVAILABLE was used only where LMI_LACKING_RINT_OR_EQUIVALENT
was conditionally (un)defined. LMI_LACKING_RINT_OR_EQUIVALENT was used
only in the negative sense:
  #ifndef LMI_LACKING_RINT_OR_EQUIVALENT
where it was not the case that it was not the case that std::rint() was
available...or where MinGW was used and therefore neither of these
macros was defined.

It is less confusing to use only LMI_HAVE_RINT, and to define it even
for MinGW (which has provided C99 rint() for some time, and provides
std::rint() as well now that lmi requires C++11). For now at least,
only the preprocessor conditionals have been modified, in order to
keep this commit simple.

Incidentally improved comments on preprocessor conditionals for
concinnity.