Add an enum whose enumerators are suitable arguments for fesetround()
Define an enum to represent C++11 rounding control, with enumerators
matching the names, type, and values of the macros defined in <cfenv>,
so that it is thus compatible with std::fe[gs]etround().

Likewise, define an enum to represent IEEE 754 precision control, with
enumerators matching the names, type, and values of the macros proposed
by WG14 N751/J11 if they are defined, or otherwise matching the names
proposed by WG14 N751/J11 and the values used by glibc.

In mapping these enums to macros, certain assumptions are made. Header
<cfenv> is included, and is assumed to define the four rounding-control
macros unconditionally, although the standard seems to guarantee that
only if the implementation provides working fe[gs]etround() functions.
All rounding macros are assumed to correspond to values representable
as decltype(FE_TONEAREST); likewise, decltype(FE_FLTPREC) is assumed to
be appropriate for all precision macros iff FE_FLTPREC is defined.

Rename various identifiers. The x87 control word parameters (e.g., the
precision-control values) used by msvc are specific to msvc, not to msw;
those used by x87 hardware are specific to the x87, not to IEEE 754.