Define floating-point environment functions for non-x87 platforms
This change allows the tests relying on the correct rounding mode being set to
pass on such platforms and generally simplifies the code as it replaces the
previous experiment (guarded by the never defined LMI_IEC_559) with using C99
functions.

Note that such platforms include x86, or x86-64, using SSE instead of x87.