Use standard <cfenv> macros with standard <cfenv> functions
At present, comparing an lmi enumerator to std::fegetround()'s result
was not an error, but only because the enumerators are contrived to
have the same names as the macros. However, that contrivance is quirky
and fragile, and it requires <cfenv> to be included in 'fenv_lmi.hpp',
whence it leaks broadly, making it harder to sequester <cfenv> as an
implementation detail and encourage lmi's equivalent but more robust
facilities to be used instead.