Check return values of <cfenv> functions
Failing to check return values of functions that use them to signal
failure is a poor practice. That an example in the C99 standard fails
to check feholdexcept()'s return value [F.9.6.6/2] is no excuse. That
C++11 fails to wrap this function and throw an error on failure is
lamentable.