Enable a test that fails for now; dismiss an old gcc defect (VZ)
A fix for the failing test has been proposed, but requires further
discussion.

Tests for char*, similar to these char[] tests, have also been added.

The gcc-4.1.2 defect was described here:

http://lists.nongnu.org/archive/html/lmi/2008-06/msg00010.html

value_cast.hpp:248: error: invalid operands of types 'const char (&)[2]'
         and 'int' to binary 'operator!='

"in spite of the fact that there is no operator!=() anywhere"

The diagnostic seems so absurd that it must have been a compiler defect.
Conditionalizing the test on that gcc version has become pointless
because we're now requiring gcc-4.9.1 for production.