Use C++11 default keyword instead of empty ctor/dtor bodies (VZ)
Give compiler more opportunities for optimization by explicitly marking
trivial ctor/dtor implementations as being default.

Changes in this commit were mostly done by clang-tidy with its
modernize-use-default check and prettied up later with

	perl -0777 -i'' -pe 's/\n= default;/ = default;/gs'

to put "= default" on the same line, instead of on a new one and some
manual alignment fixes.

The only significant change was the removal of the default ctor of
database_impl in rate_table.cpp as it was confusing to have an empty
ctor which wasn't really equivalent to the default implementation due
to the use of const member field. Clarify the intention by leaving
only the ctor taking a path and calling it explicitly with an empty
path when necessary.

[GWC notes: Several empty assignment-operator bodies have similarly
been replaced by '= default;'. The 'rate_table.cpp' change causes a
unit-test problem--see:
  http://lists.nongnu.org/archive/html/lmi/2017-01/msg00035.html
which requires investigation.]