Avoid harmless but annoying "ar u" warnings in autotools builds
Use "cr" flags with ar instead of Automake current default of "cru" to avoid
warnings about "'u" modifier ignored" due to "D" flag being on by default in
the recent binutils versions, built with --enable-deterministic-archives, used
in the latest versions of most Linux distributions.

An alternative could be to explicitly use "U" flag to disable deterministic
builds, but the (very) modest speed benefit of "u" doesn't seem to be worth it.