Adjust "ar u" warning fix to Libtool instead of Automake
We use Libtool for creating the libraries and it turns out that it uses a
different name for the same variable, so setting ARFLAGS is not enough to
suppress the warning, but AR_FLAGS is.

This confusion is already fixed in the latest Libtool, but not in any version
of it included in Debian (even unstable) yet, so it's still worth working
around it for now.