Modernize for statements
This statement in the original:
  char const*const field = fields[j];
was already not ideal because the Levenshtein distance between the
variable names is so slight that it would be too easy to mistake one
for the other. In assert_okay_to_run_group_quote() in particular, the
simplest modernization would worsen that problem, producing
  for(auto const& cell : cells)
  for(auto const& field : fields)
so the for-range-initializers have been given longer, more descriptive
names.