Modernize for statements
Struct template xml_sequence_io is documented thus:
  /// Serialization for sequences [23.1.1].
but it was not transparently obvious whether it might be used with a
container that is not a sequence, so a static assertion more or less
to that effect has been added. Even a coarse assertion is stronger
than a mere comment.