Improve each_equal() efficiency; augment its unit test
The original implementation of the iterator overload did not return
as soon as the result was known to be false (VZ). Now it does.

Added unit tests for empty ranges.

The original (commented out) invalid unit test no longer happens to
segfault reproducibly for me, but such is UB. Lest we be tempted to
treat that particular special case as an empty range in the hope of
precluding UB, another invalid example has been added that cannot be
remediated in that way because no code we write can detect it.