Reduce dissimilarities between check() and analyze_errors()
Vectors of different sizes compare unequal; there is no need first
to assert that their sizes are equal, now that std::equal() is no
longer (weirdly) used to compare them.

The most appropriate empty default for a pointer type is 'nullptr'.

It was a misunderstanding to imagine that 'k' and 'c' were comparable
in any way: 'k' means the 'a_extra_keywords' ctor argument, which 'c'
is the expected value of linear_keyword_representation().