Belay array decay
This change sacrifices some clarity due to the typedef (which prevents
lines from wrapping) for the sake of maintainability (recent changes
demonstrate the dangers of 'char const* const* p = nullptr').

Function check() is no shorter, because the default for argument 'c'
is still "special" (but constructed of Fe rather than Pu). Of course,
'c' could be defaulted to a vector of length 'n' if 'n' were made a
non-type template parameter, but that would not be simpler.

There is no need to change 'd' from array to vector: its cardinality
'n' must be passed to check() anyway for use as an InputSequence ctor
argument; and it is never defaulted, so there is no reason to default
it to nullptr.