Require ordered intervals that partition [inception, maturity) [416]
There is no need to allow intervals to be specified other than in
increasing order, and each duration must belong to one and only one
interval. The GUI input sequence editor already appropriately imposes
these requirements, and the unit test verifies the same diagnostics
that the GUI displays as warnings.

Thus, no function to sort intervals is wanted; and it is a design
invariant that the last interval end at maturity, even though the
implementation may not yet have established that invariant fully.