Improve a member name [406]
This was an improvement:
  :%s/last_possible_duration/years_to_maturity/gc
because it's more descriptive and follows usage elsewhere. This is
not "maturity age" as was suggested in a removed comment: it is
instead the number of years until that age is reached.

This possible change:
  :%s/inforce_duration/inforce_year/gc
was considered but not applied. It would follow usage elsewhere, but
"duration" unambiguously means "time since", whereas "year" could be
confused with "calendar year".